+++ src/provider/client/index.ts
@@ -640,6 +640,7 @@ export class QualysAPIClient {
       detections: vmpc.HostDetection[];
     }>,
     options?: {
+      filters?: vmpc.ListHostDetectionsFilters;
       pagination?: { limit: number };
       // TODO make this a required argument and update tests
       onRequestError?: (pageIds: number[], err: Error) => void;
@@ -647,8 +648,16 @@ export class QualysAPIClient {
   ): Promise<void> {
     const endpoint = '/api/2.0/fo/asset/host/vm/detection/';
 
+    const filters: Record<string, string> = {};
+    if (options?.filters) {
+      for (const [k, v] of Object.entries(options.filters)) {
+        filters[k] = String(v);
+      }
+    }
+
     const fetchHostDetections = async (ids: QWebHostId[]) => {
       const params = new URLSearchParams({
+        ...filters,
         action: 'list',
         show_tags: '1',
         show_igs: '1',