+++ src/provider/client/types/was/index.ts
@@ -1,0 +1,2 @@
+export * from './listWebApps';
+export * from './listWebAppFindings';