+++ src/provider/client/types/was/listWebApps.ts
@@ -1,0 +1,50 @@
+import { ISODateString, PossibleArray } from '../util';
+import { ServiceResponse } from './serviceResponse';
+
+export type ListWebAppsFilterInputParameter =
+  | 'id'
+  | 'name'
+  | 'url'
+  | 'tags.name'
+  | 'tags.id'
+  | 'createdDate'
+  | 'updatedDate'
+  | 'isScheduled'
+  | 'isScanned'
+  | 'lastScan.status'
+  | 'lastScan.date';
+
+export type ListWebAppsFilters = Partial<
+  Record<
+    ListWebAppsFilterInputParameter,
+    string | string[] | boolean | number | number[]
+  >
+>;
+
+export type ListWebAppsPagination = { limit: number; offset?: number };
+
+export type ListWebAppsResponse = {
+  ServiceResponse?: ServiceResponse<WebAppData>;
+};
+
+export type WebAppData = {
+  WebApp?: PossibleArray<WebApp>;
+};
+
+export type WebApp = {
+  id?: number;
+  name?: string;
+  url?: string;
+  owner?: Owner;
+  tags?: Tags;
+  createdDate?: ISODateString;
+  updatedDate?: ISODateString;
+};
+
+export type Owner = {
+  id?: number;
+};
+
+export type Tags = {
+  count?: number;
+};