+++ src/provider/client/types/was/serviceResponse.ts
@@ -1,0 +1,6 @@
+export type ServiceResponse<DataType> = {
+  responseCode?: string;
+  count?: number;
+  hasMoreRecords?: boolean;
+  data?: DataType;
+};