+++ src/validateInvocation.ts
@@ -1,3 +1,4 @@
+import snakeCase from 'lodash/snakeCase';
 import { URL } from 'url';
 
 import {
@@ -5,10 +6,12 @@ import {
   IntegrationValidationError,
 } from '@jupiterone/integration-sdk-core';
 
-import { DEFAULT_SCANNED_SINCE_DAYS } from './constants';
+import {
+  DEFAULT_FINDINGS_SINCE_DAYS,
+  DEFAULT_SCANNED_SINCE_DAYS,
+} from './constants';
 import { createQualysAPIClient } from './provider';
 import { QualysIntegrationConfig } from './types';
-import { getScannedSinceDate } from './util/date';
 
 const REQUIRED_PROPERTIES = [
   'qualysUsername',
@@ -40,27 +43,60 @@ export default async function validateInvocation({
     );
   }
 
-  if (typeof config.minScannedSinceDays === 'string') {
-    const minScannedSinceDays = Number(config.minScannedSinceDays);
-    if (
-      !minScannedSinceDays &&
-      !!(config.minScannedSinceDays as string).trim()
-    ) {
-      throw new IntegrationValidationError(
-        'Invalid minScannedSinceDays: ' + config.minScannedSinceDays,
-      );
-    }
-    config.minScannedSinceDays = minScannedSinceDays;
-  }
-
-  if (!config.minScannedSinceDays) {
-    config.minScannedSinceDays = DEFAULT_SCANNED_SINCE_DAYS;
-  }
-
-  config.minScannedSinceISODate = getScannedSinceDate(
-    config.minScannedSinceDays,
+  assignPropertyAsNumberFromEnvOrConfig(
+    config,
+    'minScannedSinceDays',
+    DEFAULT_SCANNED_SINCE_DAYS,
   );
+  config.minScannedSinceISODate = isoDate(config.minScannedSinceDays);
+
+  assignPropertyAsNumberFromEnvOrConfig(
+    config,
+    'minFindingsSinceDays',
+    DEFAULT_FINDINGS_SINCE_DAYS,
+  );
+  config.minFindingsSinceISODate = isoDate(config.minFindingsSinceDays);
 
   const client = createQualysAPIClient(logger, instance.config);
   await client.verifyAuthentication();
+}
+
+function readPropertyFromEnv(propertyName: string): string | undefined {
+  const envName = snakeCase(propertyName).toUpperCase();
+  return process.env[envName];
+}
+
+function parsePropertyAsNumber(
+  propertyName: string,
+  value: string | undefined,
+): number | undefined {
+  if (!value) return undefined;
+
+  const numericValue = Number(value) || undefined;
+  if (!numericValue && !!value.trim()) {
+    throw new IntegrationValidationError(`Invalid ${propertyName}: ${value}`);
+  }
+
+  return numericValue;
+}
+
+function assignPropertyAsNumberFromEnvOrConfig(
+  config: QualysIntegrationConfig,
+  propertyName: keyof QualysIntegrationConfig,
+  defaultValue: number,
+): void {
+  const envValue = readPropertyFromEnv(propertyName);
+  const configValue = config[propertyName as string];
+  config[propertyName as string] =
+    parsePropertyAsNumber(propertyName, envValue) ||
+    parsePropertyAsNumber(propertyName, configValue) ||
+    defaultValue;
+}
+
+const MILLISECONDS_ONE_DAY = 1000 * 60 * 60 * 24;
+
+function isoDate(sinceDays: number) {
+  return new Date(Date.now() - sinceDays * MILLISECONDS_ONE_DAY)
+    .toISOString()
+    .replace(/\.\d{1,3}/, '');
 }