+++ test/config.ts
@@ -1,4 +1,7 @@
-import { DEFAULT_SCANNED_SINCE_DAYS } from '../src/constants';
+import {
+  DEFAULT_FINDINGS_SINCE_DAYS,
+  DEFAULT_SCANNED_SINCE_DAYS,
+} from '../src/constants';
 import { QualysIntegrationConfig } from '../src/types';
 
 export const config: QualysIntegrationConfig = {
@@ -8,4 +11,6 @@ export const config: QualysIntegrationConfig = {
   qualysUsername: process.env.QUALYS_USERNAME || 'upter3aw',
   minScannedSinceDays: DEFAULT_SCANNED_SINCE_DAYS,
   minScannedSinceISODate: '2020-09-11T23:00:30Z',
+  minFindingsSinceDays: DEFAULT_FINDINGS_SINCE_DAYS,
+  minFindingsSinceISODate: '2020-09-11T23:00:30Z',
 };